+++ src/components/Data/Table/components/TableHeader/TableHeader.tsx
@@ -5,20 +5,20 @@ import Sorter from "./Sorter"
 import { ColumnType, Sorting } from "../../types"
 
 type Props = {
-  fixedLastColumn?: boolean
+  lastColumnFixed?: boolean
   columns: ColumnType[]
   onChangeSorting: (sorting: Sorting) => void
   sorting?: Sorting
 }
 
-const TableHeader: React.FC<Props> = ({ columns, fixedLastColumn, onChangeSorting, sorting }) => (
+const TableHeader: React.FC<Props> = ({ columns, lastColumnFixed, onChangeSorting, sorting }) => (
   <thead>
     <tr>
       { columns.map(({ header, minWidth, sorter }, index) =>
         <StyledHeader
           key={ index }
           minWidth={ minWidth }
-          isFixed={ fixedLastColumn && index === columns.length - 1 }
+          isFixed={ lastColumnFixed && index === columns.length - 1 }
         >
           { sorter ? (
               <Sorter header={ header } sorting={ sorting } onChangeSorting={ onChangeSorting } index={ index } />